[refactor_rightBar][taier-ui]refactor rightBar
