--- taier-ui/src/components/task/envParams.tsx
@@ -1,59 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import Editor from '../editor';
-import type { IEditor, IEditorTab } from '@dtinsight/molecule/esm/model';
-import { isTaskTab } from '@/utils/is';
-
-interface IEnvParams extends Pick<IEditor, 'current'> {
-	onChange?: (tab: IEditorTab, value: string) => void;
-}
-
-export default function EnvParams({ current, onChange }: IEnvParams) {
-	const handleValueChanged = (value: string) => {
-		if (current?.tab) {
-			onChange?.(current?.tab, value);
-		}
-	};
-
-	if (!isTaskTab(current?.tab?.id)) {
-		return (
-			<div
-				style={{
-					marginTop: 10,
-					textAlign: 'center',
-				}}
-			>
-				无法获取环境参数
-			</div>
-		);
-	}
-
-	return (
-		<Editor
-			sync
-			value={current?.tab?.data.taskParams || ''}
-			language="ini"
-			options={{
-				automaticLayout: true,
-				minimap: { enabled: false },
-			}}
-			onChange={handleValueChanged}
-		/>
-	);
-}
